ci: Update CircleCI outdated orbs
This commit updates the CircleCI YAML configuration file to use the latest versions of the circleci/python and circleci/ansible orbs, and to include a step that uploads the test coverage report to Codecov. It also adds environment variables for the PostgreSQL database and the Django settings module, and updates the Ansible playbook path in the deploy job.

The build job now includes a new codecov/upload step that uploads the test coverage report to Codecov after the tests have run.

The workflows section has been updated to trigger the build job for branches named develop and release, and to trigger the deploy job for branches named main.