Add OTP model with methods for generating and verifying OTPs
The OTP model has been created with fields for the user, the OTP value, and the date/time the OTP was last updated. The model includes a create method that generates a new, unique OTP for a given user, and a verify method that checks whether a given OTP is valid for a given user.

The create method ensures that the OTP is unique and marks any existing OTPs for the user as inactive. The OTP value is a string of up to four characters.

The model also includes a get_latest_otp method that returns the most recent OTP for a given user.

The updated_at field uses the auto_now_add option to set the initial value to the current date/time when the object is created, but the editable attribute is set to True to allow the value to be updated later.