Add API views and URLs for user registration, changing password, changing profile, and changing email
This commit adds views and URLs for user registration, changing password, changing profile, and changing email in a Django Rest Framework application.

The RegistrationAPIView is used to register new users and takes in a user's email and password. The ChangePasswordAPIView is used to change a user's password and takes in a user's old and new passwords. The ChangeProfileAPIView is used to change a user's profile information, except for their email and password, and is automatically generated by the SimpleRouter. The ChangeEmailAPIView is used to change a user's email address and takes in a new email address.

All of these views are mapped to their respective URLs using the path() function, with the exception of the ChangeProfileAPIView which is generated automatically by the SimpleRouter. The router is included using the include() function so that its URLs can be included in the main URL configuration.