Add Redis cache backend to Django project
This commit adds Redis as the cache backend for the project, using the django_redis.cache.RedisCache backend. The cache is configured to use the default options and a Redis server running at redis://127.0.0.1:6379/1.

The CACHE_MIDDLEWARE_ALIAS setting is set to 'default' and CACHE_MIDDLEWARE_SECONDS is set to 300 (5 minutes)